test: router перекинуть в тесты в один файл
